Replace ContainerSpec.Arguments with ContainerTests.Arguments.
